French layout improvement suggestion
Following https://github.com/dessalines/thumb-key/issues/125
